requirements.txt
python-telegram-bot[job-queue]==22.6
httpx==0.28.1
h2==4.2.0
beautifulsoup4==4.12.3
pillow==12.1.0
telegraph==2.2.0

